doc: Document the desired REST API
V1->V2:
 - Incorporated review comments (Anthony)
 - Removed unimplemented Collections to focus initial discussion

This patch introduces a new API.md file which will serve as the official
API documentation.  It provides information about the supported
Collections and Resources and the URIs where they can be accessed.  It
also describes the properties that are available for each Resource type
and actions that can be performed upon them.  The document describes
conventions and assumptions.  I look forward to your comments so that we
can charter this API with the best possible usability and extensibility.

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>
Message-id: 1364567579-19840-1-git-send-email-agl@linux.vnet.ibm.com
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>