templates: Implement the mock model
Implement the mock operations for templates.  The VMTemplate object is
used to instantiate a Template from a set of parameters.  The os_distro
and os_version fields are used to populate any unspecified fields in the
template.  This object also can create libvirt VM xml files from the
template information.

Reviewed-by: Aline Manera <alinefm@br.ibm.com>
Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>