API: Make VM name optional
In order to enable single-click VM creation from a template, it must be
possible to create a new VM with only the template URI as a parameter.
So far, the only other required parameter is 'name'.  If omitted, we can
come up with a good name based on the template name and the list of
already defined vms.

Because the low-level code can now pick an identifier for a
newly-created Resource, the model create functions must now return the
name/ident so the Collection can produce the proper return value.

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>
Reviewed-by: Anthony Liguori <aliguori@us.ibm.com>
Message-id: 1367435564-14029-5-git-send-email-agl@linux.vnet.ibm.com