ui: Create a new VM from a Template
This patch adds a new toolbar button in the Templates section that
allows you to create a new VM from a selected Template.  To wire up the
GUI events, I had to refactor some of the event code that was specific
to the VM icon area to make it generally useful.  Once done, this patch
simply wires up a POST request to the toolbar button.

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>
Reviewed-by: Anthony Liguori <aliguori@us.ibm.com>
Message-id: 1367435564-14029-6-git-send-email-agl@linux.vnet.ibm.com