ui: Display a warning if the UI needs to be built
V1->V2:
 - Implement this in pure CSS (without jQuery dependency)

A common problem that developers encounter with the new UI is forgetting
to build it before trying to use it.  When this happens, the UI is ugly
because it's missing its javascript and CSS files.  To help counter the
problem, provide an overlay that reminds people to build if the files
are missing.  When the UI is built, CSS can be used to hide the overlay
completely.

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>
Message-id: 1370985029-9536-1-git-send-email-agl@linux.vnet.ibm.com