Bug Fix: Get correct exception info when run unittest on Python2.6
When running unittest on SLES(python2.6), found a problem:
If an exception occurs in body of 'with' statement, the 'exc_value' is a str
not a exception object, 'raise firstException' will raise a TypeError.
In this case, we will not get info of the first exception.

This is a bug on Python 2.6:
http://bugs.python.org/issue7853

Reference:
http://docs.python.org/2.7/library/stdtypes.html#context-manager-types

Signed-off-by: Bing Bu Cao <mars@linux.vnet.ibm.com>