Use #unicode instead of #encoding in Cheetah templates
In Cheetah 2.4.1, the #encoding declaration must be in the beginning of the
template file or it is ignored.
The #unicode declaration can be done in any part of the file in Cheetah 2.4.1
or newer.
So use #unicode instead of #encoding to avoid restricting the declaration
position.
The #unicode and #encoding only differ in the type of text output - #unicode
returns unicode text and #enconding returns string text.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>