screenshot: Add timeout to screenshot creation
The libvirt stream functionality is not working well in libvirt package version
0.9.6 for SLES11 SP2.
While creating the screenshot image the libvirt function hangs and does not
return blocking kimchi requests to display VMs list.

So add a verification test to identify the stream functionality is working
properly. This test consists in running the screeshot creation with a timeout.
If timeout occurs, the libvirt is taking too much time to create the screenshot
image and if it happens for 10 times the stream is disabled and a black image is
shown in the screenshot view.
If libvirt successfully creates at least one screenshot image in 10 times we
can assume libvirt stream functionality works and do not need to create the
image with timeout.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>