add js globalization method
In order to realize js globalization our goals are as follows:

 1. Messages need to be in po format for translation consistency.

 2. Construction should be done at runtime

base on the two rules as above, and html can use getText to
realize it. in order to realize number 1 goal, we have to put
the js message which needs to be translated into htmls.

so we add a new file called i18n.html.tmpl.

this file is an empty file which only add the script as follows:

'burnet.list.guest.fail.msg':
$dq$_('burnet.list.guest.fail.msg')$dq,

'burnet.list.guest.fail.msg' is the key in js files , which js use
 this key like this : i18n[burnet.list.guest.fail.msg]

 $dq$_('burnet.list.guest.fail.msg')$dq} is the value , it will be
 replaced by text of different languages in the runtime when the
server generate the html.

also the i18n.html.tmpl can be auto generated by a tool in the
build time.
Signed-off-by: zhoumeina <zhoumein@linux.vnet.ibm.com>
Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>