ui: Ensure i18n.html loads
Commit cc89b1e3eb53b2e05813e6e4fe1e5ceee517b8bf introduced a bug where
the guest list would not load properly when the application is first
loaded.  This is because the i18n.html may not have been loaded and
inserted into the DOM before guestTab.html was loaded.  If this happens,
the DOM will contain errors and not load correctly.  Use the .load()
completion callback to ensure that i18n.html is loaded before
proceeding.

Fixes https://github.com/kimchi-project/kimchi/issues/73

Signed-off-by: Adam Litke <agl@linux.vnet.ibm.com>
Message-id: 1374861991-17172-1-git-send-email-agl@linux.vnet.ibm.com