template-update: resouce support update method
add PUT method for Resouce class
implement update attribute for sub-class Template

We need a common of update method.
We should also support edit the name of resource.
If we edit the name of resource, we had better raise a 301 redirection,
that means the resource is moved permanent.

add a field 'uri_fmt' that each Resource instance needs to define that
can construct the redirect uri:
self.uri_fmt = "/Collections/%s/Collections/%s"

Then in the common redirect we can just do:
raise cherrypy.HTTPRedirect(self.uri_fmt % tuple(list(self.model_args[:-1])
                            + [ident]). 301)

In the code of kimchi, somewhere, the self.model_args is defined as
list, and somewhere it is defined tuple.
so we should change it to list: list(self.model_args[:-1])

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>