add a cachebust module to generate href with cacheBust
employ a file lastModified strategy for static content cache.
cachebust will generate the href with a cacheBust.
And the value of the cacheBust is the last modified time of the
static content file.

cachebust.href
input: the static content url,  such as: css/dev.base.css
output: url with cacheBust
        such as: css/dev.base.css?cacheBust=1376668916.87

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>
Signed-off-by: Adam King <rak@linux.vnet.ibm.com>