Fix Log-in Window Redrawn after 401 Returned
Log-in window will be cleared and re-drawn after 401 returned when
username/password is wrong. This will cause message like "username
or password is incorrect." disappear. It's because login() is also
be filtered for 401 using kimchi.requestJSON() that is intended to
filter Ajax requests for RESTful APIs after the user has logged in
successfully. In this patch, we'll use jQuery.ajax() for login().

Signed-off-by: Hongliang Wang <hlwang@linux.vnet.ibm.com>