change the login window css to make language selector in the head
The language selector in the footer of the login window, it not a
good ui design. It will reload the window after change the language.
This operation will make the username and password already exsit
gone. That doesn't make sense.

This patch change the language selector to the head.
and Change some css styles to make the login window prefessional.
Add logo to the login window.
Signed-off-by: zhoumeina <zhoumein@linux.vnet.ibm.com>