bug fix: 404 error in debugger network tab when creating new VM
An 404 Not Found error is raised when trying to create a new VM:

"NetworkError: 404 Not Found - http://localhost:8000/%7Bicon%7D"
Reload the page to get source for: http://localhost:8000/%7Bicon%7D

It is caused because the html template has not loaded the image source path
when the UI is displayed. So the browser can not find the path and raises 404
errors.
To avoid it change <ul> to <script>

Signed-off-by: Aline Manera <alinefm@br.ibm.com>