Get the appropriate Kimchi version from Git tag
Since the first Makefile version, it was assumed that the Git tags were
structured as "v[0-9]*". Recently, they were changed to "kimchi-[0-9]*" and
it broke the autoconf/automake logic, so the newest tag matching "v[0-9]*"
was set on some places, like the RPM version. This patch fixes that by
matching the correct tag "kimchi-[0-9]*".

Signed-off-by: Aline Manera <alinefm@br.ibm.com>
Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>