iso streaming: Libvirt workaround
The ISO streaming support was recently added to libvirt so the most current
libvirt version in Linux distributions do not have support for it.

To allow the Kimchi user creates templates with streamed Internet ISO verify
if the libvirt has ISO streaming support and if not uses qemu/KVM directly
through <qemu:commandline> libvirt tag.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>