Issue #227: Misunderstood labels and error messages in templates tab.
V1 -> V2: Modified the msgid of the previous messages to match to the
message string and addressed some comments from reviews.

V1: Changed labels and error messages while creating a new template to don't
misunderstand the user at the time to specify an ISO file. Also added a function
to check if the file is an ISO or not.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>