Fix session locking issue
According to http://docs.cherrypy.org/dev/progguide/sessions.html:
by default, the 'locking' mode of sessions is 'implicit', which means
the session is locked early and unlocked late. It causes that while
a request is being processed, all other requests in that session
will be blocked until the request in processing is done.

The patch changes to use the 'explicit' locking mode,  which means
cherrypy will not lock the session on each request and let application
to guarantee the consistence of session data. Currently, we only use
session for authentication, so it adds code to explictly acquireing
and releasing session lock when it touches session data.

Signed-off-by: Mark Wu <wudxw@linux.vnet.ibm.com>