Add the front-end support of tabs
1) Add the tab titles in 18n.html.tmpl

2) Edit the bookmark and reload the html

3)An exmaple of tabs config tabs.xml

4) The patch series have change the tabs files from pages to pages/
tabs so need to make tabs's makefile work, this patch is working
for this purpose.

Make po work for change tabs from pages to pages/tabs

Move the tabs from pages to pages/tabs

To ensure that the tabs files are not together with other kinds
of files, we decide to move tab files to a single folder which
only contain tabs.

This is the configuration of guests, templates, storage.
And if we want to add a new tab, we can put it into pages/tabs.
and add a item in tabs.xml.
Also, we need to add the title in i18n.html.tmpl to make it have
the right translation.

5)Add a sample for plugin.
Signed-off-by: zhoumeina <zhoumein@linux.vnet.ibm.com>
Signed-off-by: Hongliang Wang <hlwang@linux.vnet.ibm.com>