Concatenate ISO URL correctly with port number
When downloading a remote ISO, the remote URL uses the port number, which
is a Python int variable. When concatenating that value with a string,
an error occurs ("TypeError: coercing to Unicode: need string or buffer,
int found").

Convert the int value to string before concatenating it with the final
URL string.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>