Issue #240: The bridge address of nat and isolated network is invalid, fix it
issue: https://github.com/kimchi-project/kimchi/issues/240

ref: http://libvirt.org/formatnetwork.html
libvirt says:
The address attribute of IP defines will be configured on the bridge
device associated with the virtual network.

But we should let user give a subnet address instead of a normal ip
address.

In this patch, kimchi will check the address.
If the address is a subnet address, kimchi will change it, and then
pass it to libvirt.

Test it:
Create a nat or isolated network in UI.
$ sudo virsh net-dumpxml your-network

Signed-off-by: ShaoHe Feng <shaohef@linux.vnet.ibm.com>