Tests: fix test_model.ModelTests.test_debug_reports
There are three problems in test_model.ModelTests.test_debug_reports.

Problem 1
It waits just 20 seconds for debug report tool to finish. It may be too
short for a big Linux installation. This patch lengthens the timeout to
60 seconds.

Problem 2
There is a time window between
  inst.debugreports_create({'name': 'report1'})
and
  rollback.prependDefer(inst.debugreport_delete, 'report1')

If the timeout mentioned above is too short, it triggers assertion
failure and it does not get a chance to insert the undo action, so a
half generated "report1" is not deleted. This in turn leads to failure
of an earlier assertion if we run the test again. That assertion is as
follow.
  self.assertFalse('report1' in report_list)

This patch put the undo inserting code right after the
debugreports_create() invocation to minimize this time window.

Problem 3
It is possible for the actions to occur in a order of following.
  debugreports_create('report1')
  debug report tool collecting information, no report file is generated yet
  asyn task wait timeout
  assertion failure: debug report async task is not finished
  rollback runs, inst.debugreport_delete('report1') is called
  current test is over, next test is run
  the debug report tool subprocess is not killed, it continues to run
    and generates a file named report1
This is possible if the system is big slow. The result is that a leftover
report file is created after the test finishes. It also triggers an
earlier assertion error mentioned above if you run "make check" twice.
The correct way is to kill the async taks and related subprocess before we
exit the test, but it requires a lot of changes in the underlying facility.
This patch just deletes the report file explicitly before running the test.

Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>