bug fix #254: Reset guest statistics when vm is powered off
The guests statistics are collected in a separated thread and it has its own
time frame to collect data.
If a vm is powered off during the sleep time frame, all the requests to the vm
will get incorrect data (previous data when vm was running).
So reset vm stats when powering off.

Signed-off-by: Aline Manera <alinefm@br.ibm.com>